refactor: split `verb.rs` out of `dict_word_metadata.rs`
